+++ include/H2DE/surfaces/H2DE_surface.h
@@ -35,50 +35,8 @@ public:
     friend class H2DE_TextObject;
 };
 
-
-
-class H2DE_Texture : public H2DE_Surface {
-private:
-    H2DE_TextureData ted;
-
-    H2DE_Texture(H2DE_Engine* engine, const H2DE_SurfaceData& sd, const H2DE_TextureData& ted);
-    ~H2DE_Texture() override;
-
-    std::optional<H2DE_AbsRect> getSrcRect() const override;
-    
-public:
-    friend H2DE_Surface* H2DE_CreateTexture(H2DE_Engine* engine, const H2DE_SurfaceData& surfaceData, const H2DE_TextureData& textureData);
-    friend void H2DE_SetTextureSrcRect(H2DE_Surface* texture, const std::optional<H2DE_AbsRect>& srcRect);
-};
-
 H2DE_Surface* H2DE_CreateTexture(H2DE_Engine* engine, const H2DE_SurfaceData& surfaceData, const H2DE_TextureData& textureData);
 void H2DE_SetTextureSrcRect(H2DE_Surface* texture, const std::optional<H2DE_AbsRect>& srcRect);
-
-
-
-class H2DE_Sprite : public H2DE_Surface {
-private:
-    H2DE_SpriteData spd;
-
-    int delayId = -1;
-    unsigned int currentFrame = 0;
-
-    H2DE_Sprite(H2DE_Engine* engine, const H2DE_SurfaceData& sd, const H2DE_SpriteData& spd);
-    ~H2DE_Sprite() override;
-
-    void initDelay();
-    void nextFrame();
-    std::optional<H2DE_AbsRect> getSrcRect() const override;
-        
-public:
-    friend H2DE_Surface* H2DE_CreateSprite(H2DE_Engine* engine, const H2DE_SurfaceData& surfaceData, const H2DE_SpriteData& spriteData);
-    friend void H2DE_SetSpriteStartingPos(H2DE_Surface* sprite, const H2DE_AbsPos& startingPos);
-    friend void H2DE_SetSpriteSize(H2DE_Surface* sprite, const H2DE_AbsSize& size);
-    friend void H2DE_SetSpriteSpacing(H2DE_Surface* sprite, int spacing);
-    friend void H2DE_SetSpriteNbFrame(H2DE_Surface* sprite, unsigned int nbFrame);
-    friend void H2DE_SetSpriteDelay(H2DE_Surface* sprite, unsigned int delay);
-    friend void H2DE_SetSpritePauseSensitive(H2DE_Surface* sprite, bool pauseSensitive);
-};
 
 H2DE_Surface* H2DE_CreateSprite(H2DE_Engine* engine, const H2DE_SurfaceData& surfaceData, const H2DE_SpriteData& spriteData);
 void H2DE_SetSpriteStartingPos(H2DE_Surface* sprite, const H2DE_AbsPos& startingPos);